inconsistency with ray angle and visual representation fixed. still issues with intersections not being checked

+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: jlehtone <jlehtone@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/10 17:07:57 by eberkowi          #+#    #+#             */
-/*   Updated: 2025/01/22 11:25:17 by jlehtone         ###   ########.fr       */
+/*   Updated: 2025/01/22 11:31:59 by jlehtone         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -98,7 +98,7 @@ typedef struct s_game
 	int		ceiling_B;
 	int		found_floor_rgb;
 	int		found_ceiling_rgb;
-	int		starting_direction;
+	float	starting_direction;
 	mlx_t				*mlx;
 	mlx_texture_t		*wall_texture;
 	mlx_image_t			*wall_img;
